requirements.txt
amqp==5.3.1
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
CacheControl==0.14.4
cachetools==6.2.2
cbor2==5.8.0
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.2.1
Django==5.2.9
django-celery-beat==2.8.1
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-filter==25.2
django-js-asset==3.1.2
django-model-utils==5.0.0
django-oauth-toolkit==3.1.0
django-redis==6.0.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
drf-yasg==1.21.11
exceptiongroup==1.3.1
firebase_admin==7.1.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jwcrypto==1.5.6
kombu==5.6.1
Markdown==3.10
msgpack==1.1.2
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.2
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.4
stripe==14.1.0
Twisted==25.5.0
txaio==25.12.2
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.1.1
